Update README.md
MomsBudgetMagic is a supportive community group for mothers. We welcome you to join us and share your tips, tricks, and advice on any topic, including mental health, self-care, budgeting, financial wellness, parenting, and anything else you might need or ask about. We are a non-judgmental group, and we look forward to getting to know you. We know that being a mother can be challenging, and we want to help you navigate the ups and downs of motherhood. We are here to support you on your journey, and we look forward to learning from you and growing together.

Signed-off-by: DesHaney6301 <destinyhaney8@gmail.com>